Moved RDDLayoutMergeMethods functionality to object.
